requirements.txt
Django==2.2
django-admin-list-filter-dropdown==1.0.3
djangorestframework==3.12.2
flake8==3.8.4
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
sqlparse==0.4.1
tqdm==4.55.0
